+++ src/modules/competition-bracket/builder/rounds-builder.tsx
@@ -1,35 +1,33 @@
-import { signal } from '@preact/signals-react';
 import { produce } from 'immer';
 import { BuilderRoundDetail } from '../types/builder-types';
-
-export const roundsStore = signal<BuilderRoundDetail[]>([]);
+import { bracketStore } from './bracket-store';
 
 export function addNewRound() {
   const roundData: BuilderRoundDetail = {
-    name: `Round ${roundsStore.value.length + 1}`,
+    name: `Round ${bracketStore.value.length + 1}`,
     matchGeneratorType: 'spot_filler',
     matchGeneratorData: {
       matches: [],
     },
   };
 
-  roundsStore.value = produce(roundsStore.value, draft => {
+  bracketStore.value = produce(bracketStore.value, draft => {
     draft.push(roundData);
   });
 }
 
 export function removeRound(index: number) {
-  roundsStore.value = roundsStore.value.filter(item => item !== roundsStore.value[index]);
+  bracketStore.value = bracketStore.value.filter(item => item !== bracketStore.value[index]);
 }
 
 export function changeRoundName(index: number, name: string) {
-  roundsStore.value = produce(roundsStore.value, draft => {
+  bracketStore.value = produce(bracketStore.value, draft => {
     draft[index].name = name;
   });
 }
 
 export function clearMatchesInRound(index: number) {
-  roundsStore.value = produce(roundsStore.value, draft => {
+  bracketStore.value = produce(bracketStore.value, draft => {
     draft[index].matchGeneratorData.matches = [];
   });
 }
