+++ src/modules/competition-bracket/types/store-types.ts
@@ -1,0 +1,6 @@
+import { AllParticipant } from './builder-types';
+
+export type SpotTypeSelections = {
+  label: string;
+  value: AllParticipant['spotType'];
+};
