+++ src/pages/competition-bracket.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
 import { NextSeo } from 'next-seo';
 import { DefaultLayout } from '~/components/default-layout';
-import { CompetitionBracketBuilder } from '~/modules/competition-bracket/competition-bracket-builder';
+import { BracketBuilder } from '~/modules/competition-bracket/bracket-builder';
 
 export default function CompetitionBracketPage() {
   return (
     <>
       <NextSeo title="Competition bracket" openGraph={{ title: 'Competition bracket' }} />
       <DefaultLayout>
-        <CompetitionBracketBuilder />
+        <BracketBuilder />
       </DefaultLayout>
     </>
   );
